Add exec to variable assignement

+++ src/tree/tree_eval.h
@@ -1,6 +1,6 @@
 #ifndef __NODE_EVAL_H__
 #define __NODE_EVAL_H__
 
-int eval(Node *node);
+double eval(Node *node);
 
 #endif /* end of include guard: __NODE_EVAL_H__ */
